Implemented the Address Object using self-reference key

+++ src/main/java/com/med/MedConnect/Model/Item/ItemRepo.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ItemRepo extends JpaRepository<Item, Long> {
+public interface ItemRepo extends JpaRepository<Item, Integer> {
 
 }
